[Lagrangian.Solver] Move MechanicalGetConstraintResolutionVisitor in its own files
